Implementado modulo de Noticias

+++ Ong.API/Ong.Domain/Command/Parceiros/DeleteParceiro/DeleteParceiroCommandHandler.cs
@@ -4,9 +4,9 @@ using Microsoft.Extensions.Logging;
 using Ong.Domain.Entities;
 using Ong.Domain.Interfaces;
 
-namespace Ong.Domain.Command.DeleteParceiro
+namespace Ong.Domain.Command.Parceiros.DeleteParceiro
 {
-    public class DeleteParceiroCommandHandler : IRequestHandler<DeleteParceiroCommand, ParceirosOng>
+    public class DeleteParceiroCommandHandler : IRequestHandler<DeleteParceiroCommand, Unit>
     {
         private readonly ILogger _logger;
         private readonly IParceirosRepository _parceirosRepository;
@@ -19,25 +19,29 @@ namespace Ong.Domain.Command.DeleteParceiro
             _parceirosRepository = parceirosRepository;
         }
 
-        public async Task<ParceirosOng> Handle(DeleteParceiroCommand request, CancellationToken cancellationToken)
+        public async Task<Unit> Handle(DeleteParceiroCommand request, CancellationToken cancellationToken)
         {
             try
             {
+                _logger.LogInformation($"Iniciado método {nameof(DeleteParceiroCommandHandler)} || Delete parceiro: {request.NomeParceiro}");
+
                 var parceiroDelete = await FindParceiroToDelete(request);
-                var result = await _parceirosRepository.DeleteAsync(parceiroDelete);
-                return result;
+                if (parceiroDelete == null) throw new ArgumentException($"Parceiro {request.NomeParceiro} não localziado");
+                await _parceirosRepository.DeleteAsync(parceiroDelete);
+
+                _logger.LogInformation($"Sucesso método {nameof(DeleteParceiroCommandHandler)} || Delete parceiro: {request.NomeParceiro}");
+
+                return Unit.Value;
             }
             catch (Exception)
             {
-                throw;
-            }
-            finally
-            {
+                _logger.LogError($"Erro método {nameof(DeleteParceiroCommandHandler)} || Delete parceiro: {request.NomeParceiro}");
 
+                throw;
             }
         }
 
-        private async Task<ParceirosOng> FindParceiroToDelete(DeleteParceiroCommand request)
+        private async Task<ParceiroOng> FindParceiroToDelete(DeleteParceiroCommand request)
         {
             var parceiroDelete = await _parceirosRepository.GetParceiroByName(request.NomeParceiro);
             return parceiroDelete is null ? throw new ArgumentException("Parceiro não localizado") : parceiroDelete;
